+++ app/src/main/java/de/baumann/sieben/workouts/MainActivity2.java
@@ -17,6 +17,7 @@ import android.widget.TextView;
 import at.juggle.sieben.SoundPool;
 import de.baumann.sieben.MainActivity;
 import de.baumann.sieben.helper.CountDownTimer;
+import de.baumann.sieben.helper.DailyStatsHelper;
 import de.baumann.sieben.pause.Pause10;
 import de.baumann.sieben.pause.Pause11;
 import de.baumann.sieben.pause.Pause2;
@@ -305,6 +306,9 @@ public class MainActivity2 extends AppCompatActivity {
 
         sharedPref.edit().putInt("ex2_number", (sharedPref.getInt("ex2_number", 0) + 1)).apply();
         sharedPref.edit().putInt("ex2_time", (sharedPref.getInt("ex2_time", 0) + (duration * 1000))).apply();
+        
+        // Track daily statistics
+        DailyStatsHelper.incrementTodayCount(this);
 
         if (sharedPref.getBoolean ("beep", false)){
             SoundPool.playWhistle(getApplicationContext());